feat: Add sign in functionality to SignIn component
This commit adds the functionality to sign in an existing account in the SignIn component. It utilizes the `signInWithEmailAndPassword` method from the Firebase auth library to authenticate the user. If successful, the user object is logged to the console. If there is an error, an alert is displayed with the error message. The sign up link now has a hover effect and underline.